% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_graph_sce.R
\name{annotate_C_emb}
\alias{annotate_C_emb}
\title{propagate a colData variable to cell embedding; endomorphism}
\usage{
annotate_C_emb(sscr, orig_var)
}
\arguments{
\item{sscr}{instance of simba_scrna}

\item{orig_var}{character(1) must be element of `colData(sscr$ppCG)`}
}
\value{
updated simba_scrna instance with `C_emb` colData updated
}
\description{
propagate a colData variable to cell embedding; endomorphism
}
