% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_graph_sce.R
\name{gen_graph_sce}
\alias{gen_graph_sce}
\title{produce graph underlying Simba PBG embedding for an sc-RNA-seq SCE (or h5ad)}
\usage{
gen_graph_sce(h5adpath, simconf = simba_config(gen_graph_copy = FALSE))
}
\arguments{
\item{h5adpath}{character(1) path to h5ad}

\item{simconf}{output of `simba_config`}
}
\value{
list with elements gdf (a data.frame of source, relation, destination, only
produced if gen_graph_copy is TRUE in simba_config), work_contents (full pathnames
of pbg outputs), workdir (a copy of the simconf$workdir)
}
\description{
produce graph underlying Simba PBG embedding for an sc-RNA-seq SCE (or h5ad)
}
\note{
simba module is imported with convert=FALSE to help zellkonverter.  In
the example gout$gdf is a python reference that could go stale.
}
\examples{
h5adpath = "/home/vincent/tenx3k.h5ad"
gout = gen_graph_sce(h5adpath, simconf = simba_config(gen_graph_copy=TRUE)) # generally want copy to be FALSE
gout$gdf
dir(gout$work_contents, full.names=TRUE, recursive=TRUE)
}
