% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_graph_sce.R
\name{joint_emb_CG}
\alias{joint_emb_CG}
\title{mutual embedding}
\usage{
joint_emb_CG(sscr)
}
\description{
mutual embedding
}
\examples{
g3k = get_fitted_3k()
jj = joint_emb_CG( g3k )
viz_joint_umap( jj, "celltype" )
}
