% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_graph.R
\name{pbg_train_config}
\alias{pbg_train_config}
\title{configuration list for training}
\usage{
pbg_train_config(
  entity_path = "",
  edge_paths = "",
  checkpoint_path = "",
  entities = structure(list(), names = character(0)),
  relations = list(),
  dynamic_relations = FALSE,
  dimension = 50L,
  global_emb = FALSE,
  comparator = "dot",
  num_epochs = 10L,
  workers = 4L,
  num_batch_negs = 50L,
  num_uniform_negs = 50L,
  loss_fn = "softmax",
  lr = 0.1,
  early_stopping = FALSE,
  regularization_coef = 0,
  wd = 0,
  wd_interval = 50L,
  eval_fraction = 0.05,
  eval_num_batch_negs = 50L,
  eval_num_uniform_negs = 50L,
  checkpoint_preservation_interval = NULL
)
}
\description{
configuration list for training
}
