% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_graph.R
\name{pbg_train_embedding}
\alias{pbg_train_embedding}
\title{after graph produced, train for embedding using PBG}
\usage{
pbg_train_embedding(
  output = "model",
  auto_wd = TRUE,
  save_wd = TRUE,
  simconf = simba_config(),
  pbg_config = pbg_train_config()
)
}
\arguments{
\item{output}{character(1)}

\item{auto_wd}{logical(1)}

\item{save_wd}{logical(1)}

\item{simconf}{like result to call to `simba_config()`}

\item{pbg_config}{list(), eventually to be handled as dict}
}
\description{
after graph produced, train for embedding using PBG
}
\examples{
h5path = "/home/vincent/tenx3k.h5ad"
g1 = gen_graph_sce(h5path)
tt = pbg_train_embedding( simconf=simba_config(workdir=g1$workdir) )
}
