% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_graph_sce.R
\name{print.simba_scrna}
\alias{print.simba_scrna}
\title{print method}
\usage{
\method{print}{simba_scrna}(x, ...)
}
\arguments{
\item{x}{instance of simba_scrna}

\item{\dots}{not used}
}
\description{
print method
}
