% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{simba_config}
\alias{simba_config}
\title{assemble some constants for a run}
\usage{
simba_config(
  min_n_genes = NULL,
  norm_method = "lib_size",
  n_top_genes = NULL,
  disc_n_bins = 5L,
  gen_graph_copy = FALSE,
  gen_graph_hvg = FALSE,
  gen_graph_dirname = "graph0",
  train_auto_wd = TRUE,
  train_save_wd = TRUE,
  train_output = "model",
  train_wd_interval = 10L,
  train_workers = 12L,
  train_wd = 0.015521
)
}
\description{
assemble some constants for a run
}
