% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_sce.R
\name{train_sce}
\alias{train_sce}
\title{train Simba PBG embedding for an SCE (or h5ad)}
\usage{
train_sce(h5adpath, simconf = simba_config())
}
\description{
train Simba PBG embedding for an SCE (or h5ad)
}
\examples{
h5adpath = "/home/vincent/tenx3k.h5ad"
train_sce(h5adpath)
}
