% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_graph_sce.R
\name{viz_cemb_umap_ly}
\alias{viz_cemb_umap_ly}
\title{interactively visualize the cell embedding}
\usage{
viz_cemb_umap_ly(gout, colour_by, ...)
}
\arguments{
\item{gout}{instance of `simba_scrna` with annotated C_emb component}

\item{colour_by}{character(1) factor in colData of C_emb for colouring}

\item{\dots}{passed to uwot::umap}
}
\value{
a plotly instance
}
\description{
interactively visualize the cell embedding
}
\examples{
g3k = get_fitted_3k()
set.seed(1234) # reproducible UMAP layout?
viz_cemb_umap_ly(g3k, "celltype")
}
