% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_graph_sce.R
\name{viz_joint_umap}
\alias{viz_joint_umap}
\title{view mutual embedding}
\usage{
viz_joint_umap(scej, cname, ptsize = 3, ltxtsize = 30, lkeysize = 1.5, ...)
}
\arguments{
\item{scej}{a SingleCellExperiment instance produced by `joint_emb_CG`}

\item{cname}{character(1) an element of colData to be used for coloring}

\item{ptsize}{for `geom_point` size}

\item{ltxtsize}{for `legend.text`}

\item{lkeysize}{for `legend.key.size` in cm}

\item{\dots}{passed to uwot::umap}
}
\value{
a ggplot instance
}
\description{
view mutual embedding
}
