% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_graph_sce.R
\name{viz_joint_umap_ly}
\alias{viz_joint_umap_ly}
\title{interactively visualize the joint embedding}
\usage{
viz_joint_umap_ly(scej, cname, ...)
}
\arguments{
\item{scej}{a SingleCellExperiment instance produced by `joint_emb_CG`}

\item{cname}{character(1) an element of colData to be used for coloring}

\item{\dots}{passed to uwot::umap}
}
\value{
a plotly instance
}
\description{
interactively visualize the joint embedding
}
\examples{
g3k = get_fitted_3k()
jj = joint_emb_CG( g3k )
viz_joint_umap_ly(jj, "celltype")
}
