% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_graph_sce.R
\name{viz_joint_umap}
\alias{viz_joint_umap}
\title{view mutual embedding}
\usage{
viz_joint_umap(
  scej,
  cname,
  plotit = TRUE,
  ptsize = 3,
  ltxtsize = 30,
  lkeysize = 1.5,
  ...
)
}
\arguments{
\item{scej}{a SingleCellExperiment instance produced by `joint_emb_CG`}

\item{cname}{character(1) an element of colData to be used for coloring}

\item{plotit}{logical(1) if FALSE, just return umap dataframe with cname in column `fac`,
and rownames disclosing cell id or gene name}

\item{ptsize}{for `geom_point` size}

\item{ltxtsize}{for `legend.text`}

\item{lkeysize}{for `legend.key.size` in cm}

\item{\dots}{passed to uwot::umap}
}
\value{
a ggplot instance, unless plotit is FALSE, in which case a data.frame is returned
}
\description{
view mutual embedding
}
\examples{
g3k = get_fitted_3k()
jj = joint_emb_CG( g3k )
set.seed(1234)
ndf = viz_joint_umap( jj, "celltype", plotit=FALSE )
ndf["CD14",]
}
